Relocating provider to lego/providers/http/webroot
